Se agrega el caso de uso DeleteProjectUseCase y la ruta correspodiente para borrar un proyecto con la condicion de que su estado sea creado

+++ src/main/java/co/com/sofka/gestionriesgos/usercases/project/DeleteProjectUseCase.java
@@ -1,7 +1,6 @@
 package co.com.sofka.gestionriesgos.usercases.project;
 
 import co.com.sofka.gestionriesgos.repositories.ProjectRepository;
-import co.com.sofka.gestionriesgos.repositories.RiskRepository;
 import com.mongodb.Function;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
@@ -11,10 +10,10 @@ import java.util.Objects;
 
 @Service
 @Validated
-public class DeleteProjectUseCase  implements Function<String, Mono<Void>> {
+public class DeleteProjectUseCase implements Function<String, Mono<Void>> {
     private final ProjectRepository projectRepository;
 
-    public DeleteProjectUseCase( ProjectRepository projectRepository) {
+    public DeleteProjectUseCase(ProjectRepository projectRepository) {
         this.projectRepository = projectRepository;
     }
 
@@ -22,6 +21,12 @@ public class DeleteProjectUseCase  implements Function<String, Mono<Void>> {
     @Override
     public Mono<Void> apply(String id) {
         Objects.requireNonNull(id, "Id es requerido");
-        return projectRepository.deleteById(id);
+
+        return projectRepository.findById(id)
+                .flatMap(project -> {
+                    if (project.getStatus().equalsIgnoreCase("Creado")) {
+                       return projectRepository.deleteById(id);
+                    }throw new IllegalArgumentException("El proyecto no puede ser eliminado");
+                });
     }
 }
